+++ djangocms_content_expiry/admin.py
@@ -1,4 +1,5 @@
 import csv
+import datetime
 
 from django.conf.urls import url
 from django.contrib import admin
@@ -6,6 +7,7 @@ from django.contrib.contenttypes.models import ContentType
 from django.http import HttpResponse
 from django.utils.translation import ugettext_lazy as _
 
+from .conf import DEFAULT_CONTENT_EXPIRY_EXPORT_DATE_FORMAT
 from .filters import (
     AuthorFilter,
     ContentExpiryDateRangeFilter,
@@ -85,6 +87,17 @@ class ContentExpiryAdmin(admin.ModelAdmin):
             ),
         ] + super().get_urls()
 
+    def _format_export_datetime(self, date, date_format=DEFAULT_CONTENT_EXPIRY_EXPORT_DATE_FORMAT):
+        """
+        date: DateTime object
+        date_format: String, date time string format for strftime
+
+        Returns a formatted human readable date time string
+        """
+        if isinstance(date, datetime.date):
+            return date.strftime(date_format)
+        return ""
+
     def export_to_csv(self, request):
         """
         Retrieves the queryset and exports to csv format
@@ -100,7 +113,7 @@ class ContentExpiryAdmin(admin.ModelAdmin):
         for row in queryset:
             title = row.version.content
             content_type = ContentType.objects.get_for_model(row.version.content)
-            expiry_date = row.expires
+            expiry_date = self._format_export_datetime(row.expires)
             version_state = row.version.get_state_display()
             version_author = row.version.created_by
             writer.writerow([title, content_type, expiry_date, version_state, version_author])
@@ -117,20 +130,21 @@ class ContentExpiryAdmin(admin.ModelAdmin):
         search_fields = self.get_search_fields(request)
         changelist = self.get_changelist(request)
 
-        changelist_kwargs = {'request': request,
-                             'model': self.model,
-                             'list_display': list_display,
-                             'list_display_links': list_display_links,
-                             'list_filter': list_filter,
-                             'date_hierarchy': self.date_hierarchy,
-                             'search_fields': search_fields,
-                             'list_select_related': self.list_select_related,
-                             'list_per_page': self.list_per_page,
-                             'list_max_show_all': self.list_max_show_all,
-                             'list_editable': self.list_editable,
-                             'model_admin': self,
-                             'sortable_by': self.sortable_by
-                             }
+        changelist_kwargs = {
+            'request': request,
+            'model': self.model,
+            'list_display': list_display,
+            'list_display_links': list_display_links,
+            'list_filter': list_filter,
+            'date_hierarchy': self.date_hierarchy,
+            'search_fields': search_fields,
+            'list_select_related': self.list_select_related,
+            'list_per_page': self.list_per_page,
+            'list_max_show_all': self.list_max_show_all,
+            'list_editable': self.list_editable,
+            'model_admin': self,
+            'sortable_by': self.sortable_by
+        }
         cl = changelist(**changelist_kwargs)
 
         return cl.get_queryset(request)