+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 unreleased
 ==========
 * fix: Content Expiry changelist filter dates should be for the future not the past
+* fix: CSV export Dates are not formatting correctly in Excel
 
 0.0.1 (2021-10-04)
 ==================